feat: rewrite middleware to generic
