+++ e2e/pierced-react/functions/_middleware.ts
@@ -1,66 +1,50 @@
-import { FragmentGateway, getMiddleware } from 'web-fragments/gateway';
+import { FragmentGateway, getStandardMiddleware } from 'web-fragments/gateway';
 
-const getGatewayMiddleware: ((devMode: boolean) => PagesFunction) & {
-	_gatewayMiddleware?: PagesFunction;
-} = (devMode) => {
-	if (getGatewayMiddleware._gatewayMiddleware) {
-		return getGatewayMiddleware._gatewayMiddleware;
-	}
+// Initialize the FragmentGateway
+const gateway = new FragmentGateway({
+  prePiercingStyles: `<style id="fragment-piercing-styles" type="text/css">
+    fragment-host[data-piercing="true"] {
+      position: absolute;
+      z-index: 9999999999999999999999999999999;
+    }
+  </style>`,
+});
 
-	const gateway = new FragmentGateway({
-		prePiercingStyles: `<style id="fragment-piercing-styles" type="text/css">
-      fragment-host[data-piercing="true"] {
-        position: absolute;
-        z-index: 9999999999999999999999999999999;
-    
-        &.remix {
-            bottom: 16%;
-            left: 15%;
-        }
-      }
-    </style>`,
-	});
+// Register fragments
+gateway.registerFragment({
+  fragmentId: 'remix',
+  prePiercingClassNames: ['remix'],
+  routePatterns: ['/remix-page/:_*', '/_fragment/remix/:_*'],
+  endpoint: 'http://localhost:3000',
+  upstream: 'http://localhost:3000',
+  onSsrFetchError: () => ({
+    response: new Response('<p>Remix fragment not found</p>', {
+      headers: { 'content-type': 'text/html' },
+    }),
+  }),
+});
 
-	gateway.registerFragment({
-		fragmentId: 'remix',
-		prePiercingClassNames: ['remix'],
-		routePatterns: ['/remix-page/:_*', '/_fragment/remix/:_*'],
-		// Note: the pierced-react-remix-fragment has to be available on port 3000
-		endpoint: 'http://localhost:3000',
-		onSsrFetchError: () => {
-			return {
-				response: new Response(
-					"<p id='remix-fragment-not-found'><style>#remix-fragment-not-found { color: red; font-size: 2rem; }</style>Remix fragment not found</p>",
-					{ headers: [['content-type', 'text/html']] },
-				),
-			};
-		},
-	});
+gateway.registerFragment({
+  fragmentId: 'qwik',
+  prePiercingClassNames: ['qwik'],
+  routePatterns: ['/qwik-page/:_*', '/_fragment/qwik/:_*'],
+  endpoint: 'http://localhost:8123',
+  upstream: 'http://localhost:8123',
+  forwardFragmentHeaders: ['x-fragment-name'],
+  onSsrFetchError: () => ({
+    response: new Response('<p>Qwik fragment not found</p>', {
+      headers: { 'content-type': 'text/html' },
+    }),
+  }),
+});
 
-	gateway.registerFragment({
-		fragmentId: 'qwik',
-		prePiercingClassNames: ['qwik'],
-		routePatterns: ['/qwik-page/:_*', '/_fragment/qwik/:_*'],
-		// Note: the pierced-react-qwik-fragment has to be available on port 8123
-		endpoint: 'http://localhost:8123',
-		forwardFragmentHeaders: ['x-fragment-name'],
-		onSsrFetchError: () => {
-			return {
-				response: new Response(
-					"<p id='qwik-fragment-not-found'><style>#qwik-fragment-not-found { color: red; font-size: 2rem; }</style>Qwik fragment not found</p>",
-					{ headers: [['content-type', 'text/html']] },
-				),
-			};
-		},
-	});
+const middleware = getStandardMiddleware(gateway, { mode: 'production' });
 
-	getGatewayMiddleware._gatewayMiddleware = getMiddleware(gateway, {
-		mode: devMode ? 'development' : 'production',
-	}) as PagesFunction;
-	return getGatewayMiddleware._gatewayMiddleware;
-};
-
-export const onRequest: PagesFunction<{ DEV_MODE?: boolean }> = async (context) => {
-	const gatewayMiddleware = getGatewayMiddleware(!!context.env.DEV_MODE);
-	return gatewayMiddleware(context);
+// CF Pages specific handler
+export const onRequest: PagesFunction = async (context) => {
+  const { request, next } = context;
+  console.log('Incoming request', request.url);
+  
+  // run the standard middleware function
+  return middleware(request, next);
 };
