+++ packages/web-fragments/src/gateway/index.ts
@@ -1,3 +1,3 @@
 export { FragmentGateway } from './fragment-gateway';
-export type { FragmentConfig } from './fragment-gateway';
-export { getMiddleware } from './utils/middleware-utils';
+export type { FragmentConfig } from './utils/types';
+export { getStandardMiddleware } from './middleware/standard';
