--- packages/web-fragments/src/gateway/middlewares/cloudflare-pages/tsconfig.json
@@ -1,7 +1,0 @@
-{
-	"extends": "../../../../tsconfig.base.json",
-	"compilerOptions": {
-		"lib": ["ESNext"],
-		"types": ["@cloudflare/workers-types"]
-	}
-}
